# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * City AI Guide API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface AuthResult {
  accessToken: string;
  user: UserProfile;
}

export interface ProfileUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  /** @nullable */
  bio?: string | null;
  /** @minLength 8 */
  password?: string;
}

export interface AvatarInput {
  /** Base64-encoded image data URI */
  avatarData: string;
}

export interface VisitedCity {
  city: string;
  country: string;
  lat: number;
  lng: number;
}

export interface TravelStats {
  totalCities: number;
  totalCountries: number;
  totalDays: number;
  /** @nullable */
  favouriteContinent?: string | null;
  visitedCities: VisitedCity[];
}

export interface CityOption {
  city: string;
  country: string;
  countryCode: string;
}

export interface FeaturedCity {
  city: string;
  country: string;
  tagline: string;
  photoUrl: string;
}

export interface PhotoResult {
  photoUrl: string;
}

export interface PlanInput {
  city: string;
  country: string;
  /** ISO date string (YYYY-MM-DD) */
  startDate: string;
  /** ISO date string (YYYY-MM-DD) */
  endDate: string;
  /**
     * No preference | Budget traveller | Mid-range | Luxury
     * @nullable
     */
  budget?: string | null;
  /**
     * Free-text interests
     * @nullable
     */
  preferences?: string | null;
  /** Solo | Couple | Family | Group */
  travellerType: string;
}

export interface PlanSummary {
  id: number;
  shareCode: string;
  city: string;
  country: string;
  startDate: string;
  endDate: string;
  travellerType?: string;
  /** @nullable */
  budget?: string | null;
  createdAt: string;
  photoUrl: string;
}

export type DestinationHowToGetThere = {[key: string]: string};

export interface Destination {
  name: string;
  category: string;
  summary: string;
  insiderTips: string[];
  entryCost: string;
  howToGetThere?: DestinationHowToGetThere;
  bestTimeToVisit: string;
  photoUrl: string;
}

export interface PlanDay {
  dayNumber: number;
  date: string;
  destinations: Destination[];
}

export interface Hotel {
  name: string;
  neighbourhood: string;
  description: string;
  priceRange: string;
  type: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  mapsUrl?: string | null;
  /** @nullable */
  bookingUrl?: string | null;
  /** @nullable */
  airbnbUrl?: string | null;
}

export interface HotelsByTier {
  budget: Hotel[];
  midRange: Hotel[];
  luxury: Hotel[];
}

export interface Restaurant {
  name: string;
  cuisine: string;
  mustTryDish: string;
  neighbourhood: string;
  priceRange: string;
  description: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  mapsUrl?: string | null;
}

export interface MiscEvent {
  name: string;
  dateOrFrequency: string;
  description: string;
  location: string;
  isFree: boolean;
}

export interface TravelPlan {
  id: number;
  shareCode: string;
  city: string;
  country: string;
  startDate: string;
  endDate: string;
  travellerType: string;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  preferences?: string | null;
  tripSummary: string;
  photoUrl: string;
  createdAt: string;
  days: PlanDay[];
  hotels: HotelsByTier;
  restaurants: Restaurant[];
  misc: MiscEvent[];
}

export type GetCityAutocompleteParams = {
q: string;
};

export type GetCityPhotoParams = {
q: string;
};