# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * City AI Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from './userProfile';

export interface AuthResult {
  accessToken: string;
  user: UserProfile;
}