# lib/api-zod/src/generated/types/destination.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * City AI Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { DestinationHowToGetThere } from './destinationHowToGetThere';

export interface Destination {
  name: string;
  category: string;
  summary: string;
  insiderTips: string[];
  entryCost: string;
  howToGetThere?: DestinationHowToGetThere;
  bestTimeToVisit: string;
  photoUrl: string;
}