# lib/api-zod/src/generated/types/hotel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * City AI Guide API
 * OpenAPI spec version: 0.1.0
 */

export interface Hotel {
  name: string;
  neighbourhood: string;
  description: string;
  priceRange: string;
  type: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  mapsUrl?: string | null;
  /** @nullable */
  bookingUrl?: string | null;
  /** @nullable */
  airbnbUrl?: string | null;
}