# lib/api-zod/src/generated/types/hotelsByTier.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * City AI Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { Hotel } from './hotel';

export interface HotelsByTier {
  budget: Hotel[];
  midRange: Hotel[];
  luxury: Hotel[];
}