# lib/api-zod/src/generated/types/miscEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * City AI Guide API
 * OpenAPI spec version: 0.1.0
 */

export interface MiscEvent {
  name: string;
  dateOrFrequency: string;
  description: string;
  location: string;
  isFree: boolean;
}