# lib/api-zod/src/generated/types/planSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * City AI Guide API
 * OpenAPI spec version: 0.1.0
 */

export interface PlanSummary {
  id: number;
  shareCode: string;
  city: string;
  country: string;
  startDate: string;
  endDate: string;
  travellerType?: string;
  /** @nullable */
  budget?: string | null;
  createdAt: string;
  photoUrl: string;
}